test(sharing-dialog): move helper tests to helpers test suite
