+++ components/sharing-dialog/src/sharing-constants.js
@@ -1,16 +1,18 @@
 import i18n from './locales/index.js'
 
+/**
+ * Access
+ */
+
 export const ACCESS_NONE = 'ACCESS_NONE'
 export const ACCESS_VIEW_ONLY = 'ACCESS_VIEW_ONLY'
 export const ACCESS_VIEW_AND_EDIT = 'ACCESS_VIEW_AND_EDIT'
 
-const noAccess = i18n.t('No access')
-
 export const accessStrings = {
     [ACCESS_NONE]: {
-        publicDescription: () => noAccess,
-        description: () => noAccess,
-        option: noAccess,
+        publicDescription: () => i18n.t('No access'),
+        description: () => i18n.t('No access'),
+        option: i18n.t('No access'),
     },
     [ACCESS_VIEW_ONLY]: {
         publicDescription: () => i18n.t('Anyone logged in can view'),
@@ -24,13 +26,14 @@ export const accessStrings = {
     },
 }
 
+/**
+ * Sharing
+ */
+
 export const SHARE_TARGET_EXTERNAL = 'SHARE_TARGET_EXTERNAL'
 export const SHARE_TARGET_PUBLIC = 'SHARE_TARGET_PUBLIC'
 export const SHARE_TARGET_USER = 'SHARE_TARGET_USER'
 export const SHARE_TARGET_GROUP = 'SHARE_TARGET_GROUP'
-
-export const isPermanentTarget = (target) =>
-    [SHARE_TARGET_EXTERNAL, SHARE_TARGET_PUBLIC].includes(target)
 
 export const defaultSharingSettings = {
     name: '',
@@ -40,33 +43,4 @@ export const defaultSharingSettings = {
     public: ACCESS_NONE,
     groups: {},
     users: {},
-}
-
-export const convertAccessToConstant = (access) => {
-    if (typeof access === 'boolean') {
-        return access ? ACCESS_VIEW_ONLY : ACCESS_NONE
-    }
-
-    if (/^rw/.test(access)) {
-        return ACCESS_VIEW_AND_EDIT
-    } else if (/^r-/.test(access)) {
-        return ACCESS_VIEW_ONLY
-    } else if (/^--/.test(access)) {
-        return ACCESS_NONE
-    } else {
-        return defaultSharingSettings.public
-    }
-}
-
-export const convertConstantToAccess = (constant, useBoolean) => {
-    switch (constant) {
-        case ACCESS_NONE:
-            return useBoolean ? false : '--------'
-        case ACCESS_VIEW_ONLY:
-            return useBoolean ? true : 'r-------'
-        case ACCESS_VIEW_AND_EDIT:
-            return useBoolean ? true : 'rw------'
-        default:
-            return useBoolean ? false : '--------'
-    }
 }
